+++ packages/salesforcedx-vscode-core/src/commands/index.ts
@@ -155,7 +155,7 @@ export {
 export { forceCreateManifest } from './forceCreateManifest';
 export { forceOrgList } from './forceOrgList';
 export { forceOrgDelete } from './forceOrgDelete';
-export { BaseDeployExecutor } from './baseDeployCommand';
+export { BaseDeployExecutor } from './baseDeployExecutor';
 export { forceFunctionCreate } from './templates/forceFunctionCreate';
 export {
   forceFunctionContainerlessStartCommand,
